refactor: move v2v ConfigMap read one level up
