+++ src/k8s/objects/v2v/vmware/vmWareDeployment.js
@@ -1,7 +1,6 @@
-import { getKubevirtV2vVmwareContainerImage, getV2vImagePullPolicy } from '../../../../config';
 import { DeploymentModel } from '../../../../models';
 
-export const buildVmWareDeployment = ({ name, namespace, kubevirtVmwareConfigMap }) => ({
+export const buildVmWareDeployment = ({ name, namespace, image, imagePullPolicy }) => ({
   apiVersion: `${DeploymentModel.apiGroup}/${DeploymentModel.apiVersion}`,
   kind: DeploymentModel.kind,
   metadata: {
@@ -26,8 +25,8 @@ export const buildVmWareDeployment = ({ name, namespace, kubevirtVmwareConfigMap
         containers: [
           {
             name,
-            image: getKubevirtV2vVmwareContainerImage(kubevirtVmwareConfigMap),
-            imagePullPolicy: getV2vImagePullPolicy(kubevirtVmwareConfigMap),
+            image,
+            imagePullPolicy,
             command: ['kubevirt-vmware'],
             env: [
               {
