+++ src/k8s/requests/v2v/constants.js
@@ -9,7 +9,7 @@ export const CONVERSION_PROGRESS_ANNOTATION = 'v2vConversionProgress';
 export const V2VVMWARE_DEPLOYMENT_NAME = 'v2v-vmware';
 
 export const VMWARE_KUBEVIRT_VMWARE_CONFIG_MAP_NAMESPACE = 'openshift'; // along the common-templates ...
-export const VMWARE_KUBEVIRT_VMWARE_CONFIG_MAP_NAME = V2VVMWARE_DEPLOYMENT_NAME;
+export const VMWARE_KUBEVIRT_VMWARE_CONFIG_MAP_NAME = 'v2v-vmware';
 
 export const VMWARE_TO_KUBEVIRT_OS_CONFIG_MAP_NAMESPACE = 'kube-public'; // note: common-templates are in the "openshift" namespace
 // TODO: make it configurable via VMWARE_KUBEVIRT_VMWARE_CONFIG_MAP
